# Optimize merge by skipping tables without edits