10x speedup for the Nautobot database:

before:
```bash
% time dolt --prof cpu merge add-ipaddress                                                      
Updating uaimen5cq6jucq3abr85klgj0g594qc1..d3922cbd8fop4ast7ohl9mm4d3dn3632
extras_objectchange | 1 +
ipam_ipaddress      | 1 +
2 tables changed, 2 rows added(+), 0 rows modified(*), 0 rows deleted(-)
dolt --prof cpu merge add-ipaddress  83.78s user 3.75s system 567% cpu 15.433 total
```

after:
```bash
% time dolt --prof cpu merge add-ipaddress
Updating uaimen5cq6jucq3abr85klgj0g594qc1..d3922cbd8fop4ast7ohl9mm4d3dn3632
extras_objectchange | 1 +
ipam_ipaddress      | 1 +
2 tables changed, 2 rows added(+), 0 rows modified(*), 0 rows deleted(-)
dolt --prof cpu merge add-ipaddress  2.27s user 0.17s system 158% cpu 1.545 total
```